https://tools.hmcts.net/jira/browse/TPET-750 - Admin E2E test
https://tools.hmcts.net/jira/browse/TPET-751 - ReadMe file change
https://tools.hmcts.net/jira/browse/TPET-783 - Account enumeration vulnerability
https://tools.hmcts.net/jira/browse/TPET-815 - Remove SECRET_KEY

+++ tests/end_to_end_tests/journey_admin/test_e2e_journey_admin.py
@@ -8,14 +8,45 @@ from tests.end_to_end_tests.helpers.e2e_page_helpers import PageHelpers
 import tests.end_to_end_tests.journey_admin.login_and_confirmation.section as section_login_and_confirmation
 import tests.end_to_end_tests.journey_admin.applications_glimr.section as section_applications_glimr
 import tests.end_to_end_tests.journey_admin.data as data
+import tests.end_to_end_tests.journey_admin.downloaded_applications.section as section_applications_downloaded
+import tests.end_to_end_tests.journey_admin.completed_applications.section as section_applications_completed
 from grc.utils.security_code import delete_all_user_codes
+from werkzeug.security import generate_password_hash, check_password_hash
+from grc.models import db, AdminUser
 
 TEST_URL = os.getenv('ADMIN_TEST_URL', 'http://grc_admin:3001')
+def ensure_e2e_admin_user():
+    user = AdminUser.query.filter_by(email=data.EMAIL_ADDRESS).first()
+    print(f"created e2e admin user: {data.EMAIL_ADDRESS}", flush=True)
+    if user is None:
+        user = AdminUser(email=data.EMAIL_ADDRESS)
+        db.session.add(user)
+
+    user.password = generate_password_hash(data.PASSWORD)
+    user.passwordResetRequired = False
+    user.userType = "ADMIN"
+    db.session.commit()
+    print(
+        f"e2e admin user created/updated: "
+        f"email={user.email}, "
+        f"exists={user is not None}, "
+        f"userType={user.userType}, ",
+        flush=True
+    )
+    lookup_user = AdminUser.query.filter_by(email=data.EMAIL_ADDRESS.lower()).first()
+
+    print(
+        f"e2e admin lookup: "
+        f"email_is_lowercase={data.EMAIL_ADDRESS == data.EMAIL_ADDRESS.lower()}, "
+        f"lowercase_lookup_found={lookup_user is not None}, "
+        f"password_matches={check_password_hash(lookup_user.password, data.PASSWORD) if lookup_user else False}",
+        flush=True
+   )
 
 def e2e_test_prerequisites():
+    ensure_e2e_admin_user()
     # Require security code to be entered
     delete_all_user_codes(data.EMAIL_ADDRESS)
-
 
 async def run_script_for_browser(browser_type):
     browser = await browser_type.launch()
@@ -41,8 +72,15 @@ async def run_script_for_browser(browser_type):
     # ------------------------------------------------
     await section_applications_glimr.run_checks_on_section(page, asserts, helpers, admin_helpers)
 
+    # ------------------------------------------------
+    # ---- Downloaded Applications section
+    # ------------------------------------------------
+    await section_applications_downloaded.run_checks_on_downloaded_section(page, asserts, helpers, admin_helpers)
 
-
+    # ------------------------------------------------
+    # ---- Completed Applications section
+    # ------------------------------------------------
+    await section_applications_completed.run_checks_on_completed_section(page, asserts, helpers, admin_helpers)
 
 async def e2e_main():
     print("")  # Blank line to improve formatting
@@ -51,7 +89,6 @@ async def e2e_main():
         for browser_type in [p.chromium]: #, p.firefox, p.webkit]:
             await run_script_for_browser(browser_type)
 
-# def test_e2e_journey_admin(app):
-#     with app.app_context():
-#         asyncio.run(e2e_main())
-
+def test_e2e_journey_admin(app):
+    with app.app_context():
+        asyncio.run(e2e_main())
